фікс багів, додана сторінка перегляду інформації про модуль для юзера, додана можливість редагування свого курсу для викладача.

+++  secondstepproject/DistanceLessons/Controllers/TeacherController.cs
@@ -21,74 +21,123 @@ namespace DistanceLessons.Controllers
             db = new DataEntitiesManager();
         }
 
+        [HttpGet]
         public ActionResult Index()
         {
             return View();
         }
 
-
-        public ActionResult Lesson(Guid id)
+        [HttpGet]
+        public ActionResult Lesson(Guid? id)
         {
+            if ((id == null) || (!db.ExistModule((Guid)id)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
             ViewBag.ModuleId = id;
-            ViewBag.Module = db.GetModulesByID(id).Title;
-            ViewBag.CourseId = db.GetModulesByID(id).CourseId;
+            ViewBag.Module = db.GetModulesByID((Guid)id).Title;
+            ViewBag.CourseId = db.GetModulesByID((Guid)id).CourseId;
             ViewBag.IsOwner = db.GetCourse(ViewBag.CourseId).UserId == db.GetUserId(User.Identity.Name) ? true : false;
             ViewBag.IsTeacher = db.GetUserRoleId(User.Identity.Name) == db.GetRoleId("Teacher") ? true : false; ;
 
-            return View(db.GetLessonsByModule(id));
+            return View(db.GetLessonsByModule((Guid)id));
         }
 
+        [HttpGet]
         public ActionResult AllMyLesson()
         {
             return View(db.UserLessons(User.Identity.Name));
         }
 
-
-        public ActionResult AllLessonsInCourse(Guid courseId)
+        [HttpGet]
+        public ActionResult AllLessonsInCourse(Guid? courseId)
         {
+            if ((courseId == null) || (!db.ExistCourse((Guid)courseId)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
             ViewBag.CourseId = courseId;
-            ViewBag.Course = db.GetCourse(courseId).Title;
-            ViewBag.Lessons = db.GetLessonsByCourse(courseId);
+            ViewBag.Course = db.GetCourse((Guid)courseId).Title;
+            ViewBag.Lessons = db.GetLessonsByCourse((Guid)courseId);
             return View();
         }
 
-        public ActionResult Course(Guid id)
+        [HttpGet]
+        public ActionResult Course(Guid? id)
         {
+            if ((id == null) || (!db.ExistCourse((Guid)id)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
             ViewBag.CourseId = id;
-            ViewBag.Course = db.GetCourse(id).Title;
-            ViewBag.Modules = db.GetModulesByCourseId(id);
-            ViewBag.OrphanLessons = db.GetOrphanLessonsByCourse(id);
-            ViewBag.IsOwner = db.GetCourse(id).UserId == db.GetUserId(User.Identity.Name) ? true : false;
+            ViewBag.Course = db.GetCourse((Guid)id).Title;
+            ViewBag.Modules = db.GetModulesByCourseId((Guid)id);
+            ViewBag.OrphanLessons = db.GetOrphanLessonsByCourse((Guid)id);
+            ViewBag.IsOwner = db.GetCourse((Guid)id).UserId == db.GetUserId(User.Identity.Name) ? true : false;
             ViewBag.IsTeacher = db.GetUserRoleId(User.Identity.Name) == db.GetRoleId("Teacher") ? true : false;
             ViewBag.IsAdmin = db.GetUserRoleId(User.Identity.Name) == db.GetRoleId("Admin") ? true : false;
             return View();
         }
-
+        /*
+        [HttpGet]
         public ActionResult MyCourses()
         {
             return View(db.GetCoursesByTeacherId(db.GetUser(User.Identity.Name).UserId));
         }
-
+        */
+        [HttpGet]
         public ActionResult Courses()
         {
             return View(db.GetCourseList());
         }
 
+        [HttpGet]
+        public ActionResult EditCours(Guid? id)
+        {
+            if ((id == null) || (!db.ExistCourse((Guid)id)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
+            ViewBag.Categories = db.GetCategoryList();
+            ViewBag.Users = db.GetGetUserByRole("Teacher");
+            return View(db.GetCourse((Guid)id));
+        }
+
+        [HttpPost]
+        public ActionResult EditCours(Cours obj)
+        {        
+            Cours old = db.GetCourse(obj.CourseId);
+            UpdateModel(old);
+            db.Save();
+            if (User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Courses", "Admin");
+            }
+            return RedirectToAction("Courses");
+
+        }
+
+
 
         [HttpGet]
-        public ActionResult CreateModule(Guid id)
+        public ActionResult CreateModule(Guid? id)
         {
+            if ((id == null) || (!db.ExistCourse((Guid)id)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
             ViewBag.CourseId = id;
             ViewBag.DateIsGood = true;
             return View();
         }
+
         [HttpPost]
         public ActionResult CreateModule(Module obj)
         {
             ViewBag.CourseId = obj.CourseId;
             obj.ModuleId = Guid.NewGuid();
             Guid courseId = obj.CourseId;
-            bool dateOk=true;
+            bool dateOk = true;
             if (obj.DateBeginTesting > obj.DateEndTesting) dateOk = false;
             ViewBag.DateIsGood = dateOk;
             if (ModelState.IsValid && dateOk) db.AddModule(obj);
@@ -98,28 +147,46 @@ namespace DistanceLessons.Controllers
         }
 
         [HttpGet]
-        public ActionResult DetailsModule(Guid id)
+        public ActionResult DetailsModule(Guid? id)
         {
-            return View(db.GetModulesByID(id));
+            if ((id == null) || (!db.ExistModule((Guid)id)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
+            return View(db.GetModulesByID((Guid)id));
         }
 
         [HttpGet]
-        public ActionResult DetailsLesson(Guid id)
+        public ActionResult DetailsLesson(Guid? id)
         {
-            return View(db.GetLessonByID(id));
+            if ((id == null) || (!db.ExistLesson((Guid)id)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
+            return View(db.GetLessonByID((Guid)id));
         }
 
 
         [HttpGet]
-        public ActionResult CreateLesson(Guid courseId, Guid mod_id, string path)
+        public ActionResult CreateLesson(Guid? courseId, Guid? mod_id, string path)
         {
+            if ((courseId == null) || (String.IsNullOrEmpty(path)) || (!db.ExistCourse((Guid)courseId)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
+
+            if (mod_id != Guid.Empty)
+            {
+                if (!db.ExistModule((Guid)mod_id))
+                    return new NotFoundMvc.NotFoundViewResult();
+            }
             Lesson obj = new Lesson();
             obj.Text = path;
             string[] filepart = path.Split('.');
             Guid res = Guid.NewGuid();
             Guid.TryParse(filepart[0], out res);
             obj.LessonId = res;
-            obj.CourseId = courseId;
+            obj.CourseId = (Guid)courseId;
             if ((mod_id != null) && (mod_id != Guid.Empty)) obj.ModuleId = mod_id;
             ViewBag.ModuleId = mod_id;
             return View(obj);
@@ -127,7 +194,6 @@ namespace DistanceLessons.Controllers
         [HttpPost]
         public ActionResult CreateLesson(Lesson obj)
         {
-
             obj.UserId = db.GetUser(User.Identity.Name).UserId;
             if (obj.UserId == db.GetCourse(obj.CourseId).UserId) obj.IsAcceptMainTeacher = true;
             obj.Publication = DateTime.Now;
@@ -144,11 +210,14 @@ namespace DistanceLessons.Controllers
         }
 
         [HttpGet]
-
-        public ActionResult DeleteLesson(Guid id)
+        public ActionResult DeleteLesson(Guid? id)
         {
-            Lesson les = db.GetLessonByID(id);
-            db.DeleteLesson(id);
+            if ((id == null) || (!db.ExistLesson((Guid)id)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
+            Lesson les = db.GetLessonByID((Guid)id);
+            db.DeleteLesson((Guid)id);
 
             if (les.ModuleId != null)
                 return RedirectToAction("Lesson", new { id = les.ModuleId });
@@ -161,6 +230,10 @@ namespace DistanceLessons.Controllers
         [HttpGet]
         public ActionResult EditLesson(Guid id)
         {
+            if ((id == null) || (!db.ExistLesson((Guid)id)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
             ViewBag.Modules = db.GetModulesByCourseId(db.GetLessonByID(id).CourseId);
             return View(db.GetLessonByID(id));
         }
@@ -188,8 +261,13 @@ namespace DistanceLessons.Controllers
             else return RedirectToAction("Course", new { id = old.CourseId });
         }
 
+        [HttpGet]
         public ActionResult AcceptLesson(Guid id)
         {
+            if ((id == null) || (!db.ExistLesson((Guid)id)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
             Lesson old = db.GetLessonByID(id);
 
             old.IsAcceptMainTeacher = true;
@@ -205,6 +283,10 @@ namespace DistanceLessons.Controllers
         [HttpGet]
         public ActionResult DeleteModule(Guid id)
         {
+            if ((id == null) || (!db.ExistModule((Guid)id)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
             Guid course_id = db.GetModulesByID(id).CourseId;
             db.DeleteModule(id);
             return RedirectToAction("Course", new { id = course_id });
@@ -214,6 +296,10 @@ namespace DistanceLessons.Controllers
         [HttpGet]
         public ActionResult EditModule(Guid id)
         {
+            if ((id == null) || (!db.ExistModule((Guid)id)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
             ViewBag.DateIsGood = true;
             return View(db.GetModulesByID(id));
         }
@@ -226,7 +312,7 @@ namespace DistanceLessons.Controllers
             bool dateOk = true;
             if (obj.DateBeginTesting > obj.DateEndTesting) dateOk = false;
             ViewBag.DateIsGood = dateOk;
- 
+
             if (ModelState.IsValid && dateOk)
                 db.Save();
             else return View(obj);
@@ -235,24 +321,25 @@ namespace DistanceLessons.Controllers
         }
 
 
-
-        public ActionResult UploadLesson(Guid courseId, Guid moduleId, string act = "create")
+        [HttpGet]
+        public ActionResult UploadLesson(Guid? courseId, Guid? moduleId = null, string act = "create")
         {
+            if ((courseId == null) || (!db.ExistCourse((Guid)courseId)))
+            {
+                return new NotFoundMvc.NotFoundViewResult();
+            }
+            if (moduleId != null)
+            {
+                if (!db.ExistModule((Guid)moduleId))
+                    return new NotFoundMvc.NotFoundViewResult();
+            }
             ViewBag.ModuleId = moduleId;
             ViewBag.CourseId = courseId;
             ViewBag.Action = act;
             return View();
         }
 
-        public ActionResult ShowFile(string filename)
-        {
-            FileRepository file = new FileRepository();
-            FileDescription lect = new FileDescription();
-            lect = file.GetFileByName(filename);
-            string completeURL = "../Uploads/" + lect.WebPath;
-
-            return Redirect(completeURL);
-        }
+        [HttpPost]
         public ActionResult Upload()
         {
             string fileName = "";
